+++ zebra-network/src/protocol.rs
@@ -2,3 +2,4 @@
 
 pub mod codec;
 pub mod message;
+pub mod types;
