+++ zebra-network/src/protocol/message.rs
@@ -8,7 +8,8 @@ use zebra_chain::block::Block;
 use zebra_chain::{transaction::Transaction, types::BlockHeight};
 
 use crate::meta_addr::MetaAddr;
-use crate::types::*;
+
+use super::types::*;
 
 /// A Bitcoin-like network message for the Zcash protocol.
 ///
